+++ src/App.jsx
@@ -1,26 +1,30 @@
 import { useState } from 'react';
 import './App.css';
-import TodoList from './TodoList';
-import TodoForm from './TodoForm';
+import TodoList from './features/TodoList';
+import TodoForm from './features/TodoForm';
 
 function App() {
   const [todoList, setTodoList] = useState([]);
   function handleAddTodo(newTodo) {
-    const newTodoID = { id: todoList.length, title: newTodo, isCompleted: false };
+    const newTodoID = {
+      id: todoList.length,
+      title: newTodo,
+      isCompleted: false,
+    };
     setTodoList([...todoList, newTodoID]);
   }
 
-  function completeTodo(id){
+  function completeTodo(id) {
     const updatedTodos = todoList.map((todo) =>
-    todo.id === id ? {...todo, isCompleted:true}:
-       todo
-    )
-    setTodoList(updatedTodos);}
+      todo.id === id ? { ...todo, isCompleted: true } : todo
+    );
+    setTodoList(updatedTodos);
+  }
   return (
     <div>
       <h1>My To-dos</h1>
       <TodoForm onAddTodo={handleAddTodo} />
-      <TodoList todoList={todoList} onCompleteTodo={completeTodo}/>
+      <TodoList todoList={todoList} onCompleteTodo={completeTodo} />
     </div>
   );
 }
